requirements.txt
asn1crypto==0.24.0
awscli==1.16.121
boto==2.49.0
boto3==1.7.4
botocore==1.10.84
cached-property==1.5.1
certifi==2018.4.16
cffi==1.12.2
chardet==3.0.4
colorama==0.3.9
comtypes==1.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
dj-static==0.0.6
Django==2.0.5
django-buckets==0.1.25
django-cors-headers==2.5.0
django-environ==0.4.5
django-filter==2.1.0
django-money==0.14.4
django-rest-swagger==2.2.0
django-s3-upload==0.2.1
django-s3direct==1.1.2
djangorestframework==3.8.2
djangorestframework-bulk==0.2.1
djangorestframework-gis==0.14
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.1.0
docker==3.7.0
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
drf-yasg==1.14.0
drfdocs==0.0.11
futures==3.1.1
GDAL==2.3.1
h5py==2.9.0
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.4
jsonschema==2.6.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lxml==4.3.2
Markdown==3.0.1
MarkupSafe==1.0
nose==1.3.7
numpy==1.15.0
openapi-codec==1.3.2
Pillow==5.4.1
psycopg2==2.7.5
psycopg2-binary==2.7.7
py-moneyed==0.8.0
pyasn1==0.4.5
pycparser==2.19
PyJWT==1.6.4
PyPDF2==1.26.0
python-dateutil==2.8.0
python-docx==0.8.10
python-magic==0.4.15
python-magic-bin==0.4.14
pytz==2018.4
PyYAML==3.13
reportlab==3.5.13
requests==2.19.1
rsa==3.4.2
ruamel.yaml==0.15.89
s3transfer==0.1.13
scipy==1.2.1
simplejson==3.16.0
six==1.12.0
static3==0.7.0
text-to-image==0.0.5
texttable==0.9.1
unoconv==0.8.2
uritemplate==3.0.0
urllib3==1.23
Wand==0.5.1
websocket-client==0.55.0
Werkzeug==0.14.1
